# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Mini Order Management API
  description: |
    A small backend system for managing products and orders.

    - **Auth:** JWT bearer tokens (short-lived access token + rotating refresh token).
    - **Prices** are returned as strings with two decimals (e.g. `"20.00"`) to avoid floating-point rounding.
    - **Rate limits:** 60 req/min per user for the API, 5 req/min per IP for auth endpoints.
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Local development server
  - url: http://localhost:8080/api
    description: Docker (nginx)
...
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange a refresh token for a new token pair
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
                  description: The refresh token issued at login or registration.
      responses:
        '200':
          description: Refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
            refresh_token:
              type: string
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...